Feat : increaseViews, increaseLikes
조회, 좋아요할 때 views, likes +1(likes는 아직 구현 전)

+++ modu/src/main/java/com/example/modu/dto/TestElement/TestsResponseDto.java
@@ -14,8 +14,8 @@ public class TestsResponseDto {
     private String username;
     private String title;
     private String image;
-    private Long views;
-    private Long likes;
+    private int views;
+    private int likes;
     private String category;
     private List<CommentResponseDto> comments;
